chore: :truck: styles.ts 없애고 전부 컴포넌트와 디자인 합침
