+++ src/components/GithubStats/StatusSection/JobStats/RangeIcon.tsx
@@ -1,4 +1,6 @@
-import { rangeIconStyle, rangeIconWrapperStyle } from './RangeIcon.styles';
+import { css } from '@emotion/react';
+
+import { SECRET_COLORS } from '@/components/GithubStats/constants';
 
 export const RangeIcon = () => {
   return (
@@ -15,3 +17,19 @@ export const RangeIcon = () => {
     </div>
   );
 };
+const rangeIconWrapperStyle = css({
+  display: 'grid',
+  gridTemplateColumns: 'repeat(5, 10px)',
+  gridTemplateRows: 'repeat(4, 10px)',
+  height: 40,
+  padding: 5,
+  gap: 0.5,
+});
+
+const rangeIconStyle = (x: number, y: number, isFilled?: boolean) =>
+  css({
+    gridColumn: x,
+    gridRow: y,
+    backgroundColor: isFilled ? SECRET_COLORS.BUTTON_ICON : 'none',
+    border: `1px solid ${isFilled ? SECRET_COLORS.BUTTON_ICON : SECRET_COLORS.BUTTON_ICON_DIMMED}`,
+  });
