+++ src/components/GithubStats/StatusSection/NameArea/Stars.tsx
@@ -1,5 +1,8 @@
+import { css } from '@emotion/react';
+
 import { StarIcon } from '@/components/GithubStats/assets';
-import { starWrapper } from './Stars.styles';
+import { COLORS } from '@/styles/colors';
+import { ELEMENT_SHADOW } from '@/components/GithubStats/constants';
 
 interface Props {
   nOfStars: number;
@@ -14,3 +17,23 @@ export const Stars = ({ nOfStars }: Props) => {
     </div>
   );
 };
+
+const starWrapper = css({
+  display: 'flex',
+  flexDirection: 'row',
+  alignItems: 'center',
+  marginBottom: -5,
+  marginLeft: -5,
+
+  svg: {
+    width: 40,
+    height: 40,
+    transform: 'rotate(15deg)',
+    fill: COLORS.WHITE,
+    filter: `drop-shadow(${ELEMENT_SHADOW})`,
+
+    ':not(:first-child)': {
+      marginLeft: -10,
+    },
+  },
+});
