+++ src/components/GithubStats/StatusSection/NameArea/index.tsx
@@ -1,7 +1,14 @@
+/* eslint-disable camelcase */
+import { Nanum_Myeongjo } from 'next/font/google';
+import { css } from '@emotion/react';
+
 import { TrustStat } from './TrustStat';
 import { Stars } from './Stars';
 
-import { nameStyle, subNameStyle } from './NameArea.styles';
+import { COLORS } from '@/styles/colors';
+import { SECRET_SIZES } from '@/components/GithubStats/constants';
+
+const nanumMyeongjo = Nanum_Myeongjo({ subsets: ['latin'], weight: '800' });
 
 export const NameArea = () => {
   return (
@@ -13,3 +20,18 @@ export const NameArea = () => {
     </>
   );
 };
+
+const nameStyle = css({
+  fontSize: SECRET_SIZES.SIZE_NAME,
+  color: COLORS.WHITE,
+  marginBottom: 5,
+  filter: `drop-shadow(0 0 2px ${COLORS.BLACK})`,
+  fontFamily: nanumMyeongjo.style.fontFamily,
+});
+
+const subNameStyle = css({
+  fontSize: SECRET_SIZES.SIZE_SUBNAME,
+  color: COLORS.WHITE,
+  fontWeight: 700,
+  filter: `drop-shadow(0 0 2px ${COLORS.BLACK})`,
+});
