+++ src/components/GithubStats/StatusSection/NumeralStats/index.tsx
@@ -1,6 +1,11 @@
-import { FastForwardIcon, HourglassIcon, PlusIcon, ShieldIcon, SwordIcon } from '@/components/GithubStats/assets';
+import { css } from '@emotion/react';
+
 import { StatChip } from './StatChip';
-import { numeralStatsTitleStyle, numeralStatsWrapperStyle, statChipDividerStyle } from './NumeralStats.styles';
+
+import { SECRET_COLORS } from '@/components/GithubStats/constants';
+import { FastForwardIcon, HourglassIcon, PlusIcon, ShieldIcon, SwordIcon } from '@/components/GithubStats/assets';
+import { COLORS } from '@/styles/colors';
+import { SIZES } from '@/styles/sizes';
 
 export const NumeralStats = () => {
   return (
@@ -44,3 +49,46 @@ export const NumeralStats = () => {
     </div>
   );
 };
+
+const numeralStatsWrapperStyle = css({
+  marginBottom: 10,
+  display: 'flex',
+  flexDirection: 'column',
+  alignItems: 'center',
+  width: 205,
+});
+
+const numeralStatsTitleStyle = css({
+  background: COLORS.WHITE_TRANSPARENT,
+  width: '100%',
+  height: 20,
+  padding: 1,
+  marginBottom: 5,
+  borderRadius: 2,
+  display: 'flex',
+  flexDirection: 'row',
+  alignItems: 'flex-end',
+
+  svg: {
+    width: 10,
+    height: 10,
+    fill: SECRET_COLORS.BUTTON_ICON_DARK,
+    marginLeft: 5,
+  },
+
+  span: {
+    fontSize: SIZES.$FONT_ML,
+  },
+});
+
+const statChipDividerStyle = css({
+  display: 'flex',
+  flexDirection: 'row',
+  alignItems: 'center',
+  justifyContent: 'space-between',
+  width: '100%',
+
+  ':not(:last-child)': {
+    marginBottom: 5,
+  },
+});
